moved ArenaMode in own files
